--- core/src/main/java/org/extism/sdk/chicory/core/ConfigurationException.java
@@ -1,12 +1,0 @@
-package org.extism.sdk.chicory.core;
-
-public class ConfigurationException extends ExtismException {
-
-    public ConfigurationException(String message) {
-        super(message);
-    }
-
-    public ConfigurationException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
