+++ core/src/main/java/org/extism/sdk/chicory/core/ExtismConfigurationException.java
@@ -1,0 +1,12 @@
+package org.extism.sdk.chicory.core;
+
+public class ExtismConfigurationException extends ExtismException {
+
+    public ExtismConfigurationException(String message) {
+        super(message);
+    }
+
+    public ExtismConfigurationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
