--- core/src/main/java/org/extism/sdk/chicory/core/HttpHostEnv.java
@@ -1,7 +1,0 @@
-package org.extism.sdk.chicory.core;
-
-import com.dylibso.chicory.runtime.HostFunction;
-
-import java.net.URI;
-import java.util.Map;
-
